chore: split agencies to separate store module, namespaced
